+++ arch/riscv/mach-dpu/plic.c
@@ -60,7 +60,8 @@ void plic_hw_mmu_init(void)
 		set_fixmap_io(FIX_PLIC_PRIO, PLIC_PRIORITYR(0, 0) & PAGE_MASK);
 		dpu_plic_reg_base = fix_to_virt(FIX_PLIC_PRIO);
 		printf("FIXMAP: %016llx -> %016llx %016llx %016llx: PLIC\n",
-		       PLIC_REG_BASE, fix_to_virt(FIX_PLIC_PRIO),
+		       (uint64_t)PLIC_REG_BASE,
+		       fix_to_virt(FIX_PLIC_PRIO),
 		       fix_to_virt(FIX_PLIC_PEND),
 		       fix_to_virt(FIX_PLIC_EN_0_31));
 	}
@@ -68,13 +69,15 @@ void plic_hw_mmu_init(void)
 		set_fixmap_io(FIX_PLIC_CTX_1, PLIC_CTX_REG_BASE(1) & PAGE_MASK);
 		dpu_plic_ctx_base[1] = fix_to_virt(FIX_PLIC_CTX_1);
 		printf("FIXMAP: %016llx -> %016llx: PLIC CTX 1\n",
-		       PLIC_CTX_REG_BASE(1), fix_to_virt(FIX_PLIC_CTX_1));
+		       (uint64_t)PLIC_CTX_REG_BASE(1),
+		       fix_to_virt(FIX_PLIC_CTX_1));
 	}
 	if (dpu_plic_ctx_base[0] == PLIC_CTX_REG_BASE(0)) {
 		set_fixmap_io(FIX_PLIC_CTX_0, PLIC_CTX_REG_BASE(0) & PAGE_MASK);
 		dpu_plic_ctx_base[0] = fix_to_virt(FIX_PLIC_CTX_0);
 		printf("FIXMAP: %016llx -> %016llx: PLIC CTX 0\n",
-		       PLIC_CTX_REG_BASE(0), fix_to_virt(FIX_PLIC_CTX_0));
+		       (uint64_t)PLIC_CTX_REG_BASE(0),
+		       fix_to_virt(FIX_PLIC_CTX_0));
 	}
 }
 #endif
