+++ kernel/cmdline.c
@@ -140,7 +140,7 @@ void cmd_console(void)
 	console_late_init();
 	readline_register_handler(cmd_console_handler);
 	con_dbg("cmd: Command section: %016llx - %016llx\n",
-		__cmd_start, __cmd_end);
+		(uint64_t)__cmd_start, (uint64_t)__cmd_end);
 }
 #else
 #define cmd_console()		do { } while (0)
