+++ kernel/page.c
@@ -138,7 +138,8 @@ static int do_page_free(int argc, char **argv)
 	page_free_pages(page_test_ptr, nr_pages);
 	if (nr_pages < page_test_num) {
 		page_test_ptr = page_offset(page_test_ptr, nr_pages);
-		printf("alloc = %016llx\n", page_test_ptr);
+		printf("alloc = %016llx\n",
+		       (uint64_t)((uintptr_t)page_test_ptr));
 	} else
 		page_test_ptr = NULL;
 	page_test_num -= nr_pages;
@@ -159,7 +160,8 @@ static int do_page_alloc(int argc, char **argv)
 	printf("Allocating page...\n");
 	page_test_ptr = page_alloc_pages(nr_pages);
 	page_test_num = nr_pages;
-	printf("alloc = %016llx\n", page_test_ptr);
+	printf("alloc = %016llx\n",
+	       (uint64_t)((uintptr_t)page_test_ptr));
 	return 0;
 }
 
@@ -170,7 +172,7 @@ int do_page_dump(int argc, char **argv)
 	list_for_each_entry(struct page, page,
 			    &page_free_list, link) {
 		printf("%016llx - %016llx\n",
-		       ptr_to_phys(page), page->end);
+		       (uint64_t)ptr_to_phys(page), (uint64_t)page->end);
 	}
 	return 0;
 }
